c75: cyclecounter: Fix get count 64 return type
- CycleCounterP_getCount64 return type has to be
uint64_t type.
- Change return type from uint32_t to uint64_t

Fixes: SITSW-3181

+++ source/kernel/nortos/dpl/c75/CycleCounterP_c75.c
@@ -44,7 +44,7 @@
 #include "CycleCounterP_c75.h"
 
 
-uint32_t CycleCounterP_getCount64()
+uint64_t CycleCounterP_getCount64()
 {
     uint64_t count = __TSC;
     return count;
